+++ src/Akeneo/Pim/WorkOrganization/TeamworkAssistant/Component/Repository/ProjectRepositoryInterface.php
@@ -17,7 +17,7 @@ use Akeneo\Tool\Bundle\StorageUtilsBundle\Doctrine\ORM\Repository\CursorableRepo
 use Akeneo\Tool\Component\StorageUtils\Cursor\CursorInterface;
 use Akeneo\Tool\Component\StorageUtils\Repository\IdentifiableObjectRepositoryInterface;
 use Akeneo\Tool\Component\StorageUtils\Repository\SearchableRepositoryInterface;
-use Doctrine\Common\Persistence\ObjectRepository;
+use Doctrine\Persistence\ObjectRepository;
 
 /**
  * @author Arnaud Langlade <arnaud.langlade@akeneo.com>