+++ src/Akeneo/ReferenceEntity/back/Domain/Event/ReferenceEntityRecordsDeletedEvent.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Akeneo\ReferenceEntity\Domain\Event;
 
 use Akeneo\ReferenceEntity\Domain\Model\ReferenceEntity\ReferenceEntityIdentifier;
-use Symfony\Component\EventDispatcher\Event;
+use Symfony\Contracts\EventDispatcher\Event;
 
 /**
  * Event triggered when all records belonging to a reference entity are deleted from DB