+++ CHANGELOG.md
@@ -72,7 +72,10 @@
 - PIM-10020: Fix asset media link preview when the URI contains special characters
 - PIM-10018: Fix asset code case in the product edit form
 - PIM-9989: Fix record code of reference entity field is case-sensitive
+- PIM-10021: Product link rule too long to be completed
 - PIM-10026: Avoid session persistance for API
+- PIM-10028: Fix the ETA time displayed when executing a rule with several actions
+- PIM-10042: Fix infinite loop when using get all reference entity records API endpoint
 
 ## Improvements